Upgrade Rust

+++ components/layout/util.rs
@@ -123,7 +123,7 @@ impl OpaqueNodeMethods for OpaqueNode {
     fn from_thread_safe_layout_node(node: &ThreadSafeLayoutNode) -> OpaqueNode {
         unsafe {
             let abstract_node = node.get_jsmanaged();
-            let ptr: uintptr_t = abstract_node.reflector().get_jsobject() as uint;
+            let ptr: uintptr_t = abstract_node.reflector().get_jsobject() as uintptr_t;
             OpaqueNode(ptr)
         }
     }
